feat: Add AsyncStorage to retrieve user ID in HomeScreen
This commit adds the use of AsyncStorage in the HomeScreen component to retrieve the user ID. The retrieved user ID is then stored in the state variable 'user'. This change ensures that the user ID is available for further operations in the component.

Note: This commit message follows the established convention of starting with a verb in the imperative form and providing a concise description of the change.