feat: Add expo-image-picker dependency for image selection in mobile app
This commit adds the "expo-image-picker" dependency to the "mobile/package.json" file. The dependency is required for image selection functionality in the mobile app. By including this dependency, users will be able to select images from their device's gallery or camera when creating a post.

Note: This commit message follows the established convention of starting with a verb in the imperative form and providing a concise description of the change.