feat: Add refresh function to NewPostModal
This commit adds a refresh function to the NewPostModal component in order to update the UI after a new post is created. The refresh function is passed as a prop from the HomeScreen component and is called after the modal is closed. This change ensures that the new post is immediately visible to the user without having to manually refresh the page.

Note: This commit message follows the established convention of starting with a verb in the imperative form and providing a concise description of the change.