feat: Add commentService to TabLayout and HomeScreen
This commit adds the commentService dependency to the TabLayout component and the HomeScreen component. The commentService is used to handle comments related functionality, such as retrieving and posting comments. By including this dependency, users will be able to interact with comments on posts.

Note: This commit message follows the established convention of starting with a verb in the imperative form and providing a concise description of the change.