refactor: Add change picture functionality to ProfileScreen
This commit adds the functionality to change the profile picture in the ProfileScreen. It introduces two new state variables, 'changePictureModalVisible' and 'changePictureModalVisibleAvatar', which control the visibility of the change picture modals for the cover image and avatar image respectively. The modals are triggered by pressing on the cover image or avatar image, and allow the user to select a new image from their device. The selected image is then uploaded to the server and the user's profile is updated with the new image.

Note: This commit message follows the established convention of starting with a verb in the imperative form and providing a concise description of the change.